health units (retrieving+dropping)
this program is for the health units mission. This code retrieves a health unit from the dance floor, and then comes back to base. I then manually load the other health units provided in the home area while the wait command is running for 5 seconds. After the 5 seconds is over, the bot heads towards the RePlay logo and drops off the 5 collected health units and then goes back the base to complete the mission.